apps/agent/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic-settings==2.5.2
httpx==0.27.2

